Add dependancy checks for various services
Yes, I know, a big patch. This basically renames every bash command
to use a variable of the same name, prepended with an underscore. Then
each service has a dependencies file in the etc/ directory to determine
if all the commands that it needs to function exist on the system. If
not, then disable the service.