Add relay_host mail functionality
Mail sent to this server for a domain can now accept a per-domain
relay host to relay mail for. Great when you're hosting the domain
on the server but not the email, and there is no MX record.

Could also be used as a hack-n-slash SPAM email gateway if you have
amavisd-new configured.