Add 3rdparty.include file, and .local support for them
This commit creates the 3rdparty.include file, and edits the apache
conf files to reference it. Also, if a .local for an app exists,
then use that instead of the stock ravencore. This provides a relativly
easy method to upgrade/rollback these apps, or completly switch it out
(in the case of replacing the webmail with something else).