Upgrade phpMyAdmin from 2.9.2 to 2.11.9.4
I'd like to go to 3.1.x but it requires PHP 5 and there are still a lot
of poeple on PHP 4 using ravencore.

For now I'm just changing phpmyadmin to use http auth and I'm not going
to mess around with the session crap as I had before. The downside here
is you'll get prompted for the password when you click the linkg ... but
that's not a big deal :P