Add "rpm" to Makefile
Running "make rpm" will now run git2rpm.sh in such a way that
it only builds the rpm (it now needs an argument to do the
re-install).

Helpful for development snapshots in which I won't be releasing
an RPM, but people can build one to upgrade if they like without
much fuss.