Add option to allow access to webstats w/o going through control panel
With this turned on for a domain, you can go to a URL like this:

http://dev.ravencore.com/awstats/awstats.pl

And it'll bring up the webstats for that domain. All you have to do is
update the "webstats_url" field in the domains table (which defaults to
no) to "yes" and run rehash_httpd, and it'll work.

Keep in mind that this opens up webstats for anyone to view, and you can
also pass the "config" value to the URL to pull up other people's webstats
(assuming the target domain is setup for public access) like this:

http://dev.ravencore.com/awstats/awstats.pl?config=some.other.com

This might be a problem when you want to add a .htaccess rule to the config
for webstats, authenticated users on the system still can view yours. I'll
see if I can add a patch for this to awstats to remove the ability to pass
?config= to the URL.

This commit doesn't give a switch in the control panel to turn this on.
It's manual for now. I'll add this later.