Make git2rpm.sh play nice with various symlinks
I like to point /usr/local/ravencore to my git repo. This of
course causes all kinds of problems, because ravencore creates
a bunch of files when it starts up. So this patch does:

* adds all those files to .gitignore
* forces the tar command to follow symlinks
* moves LICENSE, GPL, and README.install to the right place
* takes the copy of the above files out of the Makefile
* creates symlinks to replace those files

I may find some other quirks that come up. But as it is, this
patch seems to fix the initial issues of having symlinks in
the dev/build environment.