Add reload_webserver function
This patch does two things:

1) reverts the commit that makes ravencore.httpd a symlink to
   the system's httpd binary

2) adds the reload_webserver function (which really just calls
   start_webserver with the "graceful" argument)

The symlink is qwarky. The "pidof" ran in "service httpd status"
defaults to "pidof /usr/sbin/httpd" which then thinks apache is
running, when it's not, but the control panel is. Gosh.

Anyway, the reason I symlinked it in the first place is you got
a "binary file busy" error when you updated apache. Well, if you
unlink the ravencore.httpd file then re-copy it (and tell it to
reload gracefuly) you bypass this problem. So there you have it.