Start and stop daemon and web interface seperatly
This patch changes how the init script starts things. First it
starts the daemon as normal, then connects to the socket and tells
it to start the webserver.

Also add a connect retry in RavenCore::Client (tries three times)
just in-case it takes a second to have the daemon listen on
startup.