Fix online version check
There is a version checker that runs when you login. Based on
configuration, this can lock non-admin users out when a new
release is out, which is a security-ish feature (say there was
an exploit from an authenticated user, this would help negate
it). Somewhere along the line this check broke.

This patch fixes it.