Fix username variable in init script
Probably not something anyone would really notice, but did needed to be
fixed. When you start ravencore on a new install or upgrade, it said:

web interface as the  user

instead of

web interface as the admin user

This patch fixes it. It was incorrectly referencing the username variable,
and didn't even source the file in the first place.