Split back-end functions out to varous pm files
Lets face it, the Server.pm file is WAY to large. Not that
it's a bad thing really, but c'mon.

This patch splits out the non server-esque functions out to
other files, so basically all that is left inside Server.pm
is stuff relating to the Net::Server module and other self
functions required to startup the daemon.

Database related functions were put into the
new database.pm file, auth related functions got added to
the auth.pm file, and the rest went into misc.pm

The misc.pm is still 2001 lines long, and I'll probably
split it up eventually as well. But this patch as-is makes
everything much more manageable.