Be more lax on permissions for rc.sock
At least for now, allow any user in the system to write to the
socket. You have to know a valid username and password to be able
to do anything. I do want to revert this patch eventually, but
for now this makes testing the API a little less aggravating.