Add ability for email user to change password
This patch implements the ability for the email user to
change his or her password via the following squirrelmail
plugin:

http://squirrelmail.org/plugin_view.php?id=107

It's chgsaslpasswd file is replaced by a perl script
which makes a call to the RavenCore::Client module, and
authenticates via the email user, and changes the
password.

Also, this patch adds the ability for an email user to
authenticate against the socket. So you could technically
login to the ravencore interface as the email user, but
it (currently) won't let you do anything.