Fix the IfModule in httpd.include for php
At one point in the past this worked:

<IfModule sapi_apache2.c>

But it doesn't seem to work anymore. At least, not on my
CentOS5 test machine. Anyway, I fixed this by having all
the data in it print twice, one for mod_php5 and one for
mod_php4 (since a lot of people still use it).

If you upgrade from a previous version, you may want to
double check to make sure your PHP applications still work,
because of the open_basedir restriction may have not have
been in effect for you.