Enforce "from" address in php mail() function
The mail() function in php makes use of the sendmail_path value
from the php.ini file. This sets it specifically per domain with
the -f to be the httpd_user @ the_domain, so bounces go to that
domain (rather then the hostname of the server).

This really only affects people who don't specificy the fourth
argument to the mail() function, which is: additional_headers