Change how run_query accepts input for data
Historically the run() function would just accept a string and split it
out on a space. The first element of the array would be the function to call,
and the rest passed to that function as a string.

This patch changes that so run takes two arguments:

1) the function name
2) a hash reference (or in PHP, an array)

The function is passes as plain text, and the hash reference is serialized and
base64 encoded so it can be transported just like the return values.

This allows for more complicated data to be passed from client to server, for
future desired functionality.