Delete NULL elements in run_query()
If PHP passes an Array() with keys that have an undefined value, the
serialize actually translates that to the character NULL. I'm not sure
if that's the intended behaviour, but rather then changing it I am
opting for just fixing them in translation.

This patch looks for values of character NULL and deletes them from
the hash reference.

I noticed this in the auth() function when you had no session and
you weren't posting login credentials. The username/password elements
existed, but didn't appear to contain any data. Turns out that they
were character null, which showed login failure when you didn't post
anything.