Fix regression in session_write()
With the recent changes to how functions are called, I noticed that data
wasn't being persisted as it should. The easiest way to see this is to
change the language on the login screen and see if that persists past
the login page.

This patch fixes it.

Also, make session_decode() available in Serialize.pm Exports and use
it where it needs to be.