Fix apache user check on config file when apache not running
Turns out the "1" was coming from the perl version of grep where in scalar
context, it returns the number of matches (instead of the matching line).
Changed the code to just make a shell call to the system grap command
and parse it that way.

Also pipe the ps output to /dev/null so when apache isn't running, don't
spit out the ps error output.