Check for a tty on build
I'm setting up hudson to manage ravencore builds. As it is right now, it
does a checkout and isn't on branch master when it builds, but has the
master branch checked out. So when there isn't a tty, assume that we're
being built in hudson, and skip the check because it prompts the user
and thus fails the build.