Require a commit to build new changes
Due to the change in using git-archive rather than tar,
changes to the working tree don't get picked up on
snapshot builds. This has become a problem, and since
I don't have any current way to use git-archive to copy
the working tree (my figure out something for this in
the near future), for now just refuse to build if there
are any changes to the checkout or index as a reminder
to commit them.