Fix /etc/group bug where servgrp gets mangled
Upon restarts of RavenCore, this would wind up in /etc/group:

   servgrp::501:ARRAY(0x205ea580),apache,rcadmin

This is because when the file is being read, the variable gets initialized
as a string instead of an array. This adds the check for 'user_list', and
if so, reads the data in as an array, splitting off of a comma.