Refactor redirect_addr section of rehash_mail
First off; why is this a db query in the first place? There's only
one possible entry, and we already have the data in the $mail ref,
so just make this an if instead of a while (or foreach).

I must have been spacing when I first wrote this all those years ago.