Initial commit of push_hosting function
This is taken from the PHP code that was removed a few commits ago;
basically do the dirty work for add/edit/delete of hosting for a
domain.

There are a few things missing from here, such as the permissions
for cgi/php/etc, and there may be some other regressions. I'll follow
up with fixes when I get a chance to test them.