Use error framework for login errors
Normal errors are not reported on the login screen, which makes it
incredibly difficult to debug when there are problems logging in that
have nothing to do with authentication. This changes it so 1) the
login system itself uses the do_error method to report why login failed
and 2) moves the ERRORS comment section outside of the section that
makes it only display if you've got a valid login session.