Initial commit to remove MySQL back-end support
You know what? Screw supporting more than one database backend for
RavenCore's own function. I'm opting for using the serverless engine
because this control panel doesn't need what a client/server database
has to offer.

Now that's not to say I'm removing support for creating MySQL databases
and users; and although this patch will kill that functionality, I'll
add that specific thing back in before I release. It just needs to have
a different approach if I'm going to be supporting more than just one
database engine for that purpose.

But as I said; for the purposes of the control panel's own function,
using just SQLite is fine, especially since it's in the Public Domain.