Don't build the source binary package
Since the change made to use git-archive and getsrc from the Makefile
to build RavenCore, the src.rpm hasn't been more than a few k in size
greater than the tarball, and had nothing else in it. The binary rpm
that is generated, on the other hand, is close to 8MB.

So there really is not point in publishing the .src.rpm file, since
it's made from the Makefile within the .tar.gz file anyway.