Add bare target to Makefile
This commit does a couple of things, all to make is so if "bare" is
made as a target, GPL applications don't get downloaded and built.

First add a SUFFIXES macro so we have bare.info rather than just
a "bare" file. Second, this is an "empty target" rather than a phony
target. Third, add bare.info to .gitignore but make sure clean doesn't
nuke the file. Then put the GPL stuff inside a new target, gplbuild,
and only call it at the end of dobuild if bare.info does not exist.