Use perl module Digest::SHA::PurePerl
Since the .shadow file is no longer needed to connect to a database,
and it acts as the administrator password, it would be a good idea
to NOT store that as plain text. This commit adds the Pure perl
implementation of the sha1sum to the code base so a future commit
can hash the .shadow password file.