Use hashing algorithm or storing the admin password
The .shadow file has previously been used to connect to the
database as well as dub as the administrator password. This
commit changes the password functions around to use sha1 to
encrypt the .shadow password, with a 24 character salt, since
it is no longer needed in plain text for database use.