Refactor rehash_httpd to use Template files
When building the vhosts.conf file, use template files instead of
having all the string code inline to the perl scripts. This is going
to make features easier to add, as they're just template files and
not code. As you can see, there is a lot of stuff being deleted here.

This commit removes the idea of httpd.include and vhost.conf files. In
a future commit I'll add the ability for per-domain template files.