Refactor how modules get enabled and disabled
First we have to change where the code looks for the files,
then we have to use the existance of a few files to determine
if 1) the module is installed or not, and 2) whether or not
the user want to actually use it.

This gets rid of the pesky chmod to the .conf files that kept
showing up in "git status". I can include a file in the
.gitignore, but I can't ignore just a change in permissions
on a file.