Require connect info for the mysql module
These get created for you if you upgrade from 0.3.x, but fresh
installs are not going to have any idea what connect info is,
and we're no longer invasivly taking command of the mysql server
anymore. So from now on, it's up to the administrator to make sure
that these are kept up to date if they want databases managed
via RavenCore.