Setup wildcard domains on VHOST_WILDCARD_IP
Previously, get_domains_with_no_ip would be used in a custom "wildcard"
virtualhost. While this worked, it doesn't work when a domain is pointed
to an IP address that is set up for IP hosting, and it isn't defined on
that IP address. You end up with that domain showing up as whatever the
default domain is for that IP.

You are now forced to set a VHOST_WILDCARD_IP when you install ravencore,
and any domain with no IP gets setup on this IP address.