Reorganize lib into server and common
First, move server/var/lib as just lib ... since it doesn't make
a whole lot of sense to have lib inside var. This also keeps
the lib directory's location in sync with the new common directory,
which lib and sbin contain files that will be in use by both
the server and the (soon to come) slave.

This commit just moves files around. More to come to fix what
this breaks.