Replace module files with single settings.ini
This removes an assortment of files in favor of ini format, to keep things
in one place and far simpler. Each setting has an INFO for that module which
contains dependencies, etc, and each other section is a configuration setting,
where we define the description, the type (so it can be validated against a
function), and the default vaule.